Feat/#5 mypage screen구현

+++ feature/mypage/src/main/java/com/umcspot/spot/mypage/recruiting/navigation/RecruitingStudyNavigation.kt
@@ -18,7 +18,7 @@ fun NavGraphBuilder.myRecruitingStudyGraph(
     onRegisterScrollToTop: ((() -> Unit)?) -> Unit,
     onStudyClick : (Long) -> Unit,
     moveToMakeStudy :() -> Unit,
-    moveToCheckApplied: () -> Unit
+    moveToCheckApplied: (Long) -> Unit
 ) {
     composable<MyRecruitingStudy> {
         RecruitingStudyScreen(
